Correct tracing in UDP and serial
