+++ pyuavcan/transport/_tracer.py
@@ -4,6 +4,7 @@
 
 from __future__ import annotations
 import abc
+import enum
 import typing
 import dataclasses
 import pyuavcan
@@ -18,6 +19,21 @@ class Capture:
     for them such that it is always possible to determine which transport an event has arrived from using a single
     instance check.
     """
+
+    class Direction(enum.Enum):
+        """
+        Differentiates the initiator of the captured event: either the local node or an external entity.
+        """
+
+        RX = enum.auto()
+        """
+        Event is triggered by an external entity. When sniffing on a bus, all frames are marked as RX fragments.
+        """
+        TX = enum.auto()
+        """
+        This is rather uncommon, it represents the case where the capturing node is also engaged in network exchange.
+        Typically, a capturing unit would remain silent, so all captures would be RX captures.
+        """
 
     timestamp: pyuavcan.transport.Timestamp
